+++ SoftwareCenter.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """SoftwareCenter - Desktop-Organizer für Software-Verknüpfungen."""
 
-__version__ = "1.0.0"
+__version__ = "1.2.0"
 
 import configparser
 import json